refactor(mentor): components, handleLogoChange 오타수정
